0010: perf. threads
