--- apps/threads/src/config-factory/config-factory.model.ts
@@ -1,16 +1,0 @@
-export type AppConfig = {
-  port: number;
-  nodeEnv: string;
-  address: string;
-  printConfig: boolean;
-  bodyLimit: number;
-  enableSwagger: boolean;
-  cors?: {
-    origin?: string;
-    methods?: string;
-    preflightContinue?: boolean;
-    optionsSuccessStatus?: number;
-    credentials?: boolean;
-    allowedHeaders?: string;
-  };
-};
