--- apps/threads/src/constants/socket-io.constants.ts
@@ -1,5 +1,0 @@
-export const SOCKET_IO_EVENT = Object.freeze({
-  TOPIC: 'TOPIC',
-  THREAD: 'THREAD',
-  MESSAGE: 'MESSAGE',
-} as const);
