--- apps/threads/src/errors/emit-event.error.ts
@@ -1,6 +1,0 @@
-export class EmitEventError extends Error {
-  constructor(message?: string, cause?: unknown) {
-    super(message, { cause });
-    this.name = this.constructor.name;
-  }
-}
