+++ apps/threads/src/main.module.ts
@@ -1,39 +1,50 @@
-import { BullMQModule } from '@ckir.io/bullmq';
+import {
+  BULLMQ_QUEUE,
+  BullMQModule,
+  BullMQPinoLoggerModule,
+} from '@ckir.io/bullmq';
+import { ConfigFactoryModule } from '@ckir.io/config-factory';
 import { SocketIOModule } from '@ckir.io/socket-io';
 import { Logger, Module } from '@nestjs/common';
 
-import { ConfigFactoryModule } from './config-factory/config-factory.module';
-import { ConfigFactoryService } from './config-factory/config-factory.service';
-import { BULLMQ_QUEUE } from './constants/bullmq.constants';
+import { AppConfigService } from './configs/app-config.service';
+import { BullMQConfigService } from './configs/bullmq-config.service';
+import { SocketIOConfigService } from './configs/socket-io-config.service';
 import { ThreadsController } from './controllers/threads.controller';
 import { ThreadProcessor } from './processors/thread.processor';
 import { ThreadsService } from './services/threads.service';
 
 @Module({
+  controllers: [ThreadsController],
+  providers: [Logger, ThreadsService],
   imports: [
-    ConfigFactoryModule.forRoot({ global: true }),
+    ConfigFactoryModule.forRoot({
+      global: true,
+      providers: [AppConfigService, BullMQConfigService, SocketIOConfigService],
+    }),
+    BullMQPinoLoggerModule.registerAsync({
+      inject: [BullMQConfigService],
+      useFactory: async ({ pinoConfig }: BullMQConfigService) => pinoConfig,
+    }),
     BullMQModule.registerAsync({
       global: true,
-      inject: [ConfigFactoryService],
+      inject: [BullMQConfigService],
       queues: [
         BULLMQ_QUEUE.BROADCAST_THREAD,
         BULLMQ_QUEUE.PERSIST_THREAD,
         BULLMQ_QUEUE.VECTORIZE_THREAD,
       ],
       processors: [ThreadProcessor],
-      usePinoFactory: async ({ pinoConfig }: ConfigFactoryService) =>
-        pinoConfig,
-      useBullFactory: async ({ bullMQConfig }: ConfigFactoryService) =>
+      usePinoFactory: async ({ pinoConfig }: BullMQConfigService) => pinoConfig,
+      useBullFactory: async ({ bullMQConfig }: BullMQConfigService) =>
         bullMQConfig,
     }),
     SocketIOModule.registerAsync({
       global: true,
-      inject: [ConfigFactoryService],
-      useFactory: async ({ socketIOConfig }: ConfigFactoryService) =>
+      inject: [SocketIOConfigService],
+      useFactory: async ({ socketIOConfig }: SocketIOConfigService) =>
         socketIOConfig,
     }),
   ],
-  providers: [Logger, ThreadsService],
-  controllers: [ThreadsController],
 })
 export class MainModule {}
