+++ apps/threads/src/main.ts
@@ -1,13 +1,3 @@
-import compress from '@fastify/compress';
-import { Logger, VersioningType } from '@nestjs/common';
-import { NestFactory } from '@nestjs/core';
-import {
-  FastifyAdapter,
-  NestFastifyApplication,
-} from '@nestjs/platform-fastify';
-import { SwaggerModule } from '@nestjs/swagger';
-
-import { ConfigFactoryService } from './config-factory/config-factory.service';
 import {
   API_DOCS,
   BODY_LIMIT,
@@ -17,7 +7,17 @@ import {
   logSwaggerPath,
   SWAGGER_DOCUMENT,
   VALIDATION_PIPE,
-} from './constants/main.constants';
+} from '@ckir.io/helpers';
+import compress from '@fastify/compress';
+import { Logger, VersioningType } from '@nestjs/common';
+import { NestFactory } from '@nestjs/core';
+import {
+  FastifyAdapter,
+  NestFastifyApplication,
+} from '@nestjs/platform-fastify';
+import { SwaggerModule } from '@nestjs/swagger';
+
+import { AppConfigService } from './configs/app-config.service';
 import { MainModule } from './main.module';
 
 void (async () => {
@@ -28,13 +28,13 @@ void (async () => {
     adapter,
     logger,
   );
-  const factory = APP.get(ConfigFactoryService);
+  const appConfigService = APP.get(AppConfigService);
   await APP.register(compress as any, {
     threshold: 1024, // minimum payload size to compress
     encodings: ['br', 'gzip'], // optional: restrict Brotli/gzip
     global: true, // default behavior – compress all
   });
-  APP.enableCors(factory.appConfig.cors);
+  APP.enableCors(appConfigService.appConfig.cors);
   APP.enableVersioning({
     type: VersioningType.URI,
     defaultVersion: '1',
@@ -49,14 +49,14 @@ void (async () => {
   );
   await APP.listen(
     {
-      port: factory.appConfig.port,
-      host: factory.appConfig.address,
+      port: appConfigService.appConfig.port,
+      host: appConfigService.appConfig.address,
     },
     () => {
       const nestLogger = APP.get(Logger);
-      logConfigObject(nestLogger, factory);
-      logServerPath(nestLogger, factory);
-      logSwaggerPath(nestLogger, factory);
+      logConfigObject(nestLogger, appConfigService.appConfig);
+      logServerPath(nestLogger, appConfigService.appConfig);
+      logSwaggerPath(nestLogger, appConfigService.appConfig);
     },
   );
 })();
