+++ libs/bullmq/src/bullmq/bullmq.constants.ts
@@ -9,5 +9,7 @@ export const BULLMQ_QUEUE = Object.freeze({
   PERSIST_TOPIC: 'PERSIST_TOPIC',
   PERSIST_THREAD: 'PERSIST_THREAD',
   BROADCAST_TOPIC: 'BROADCAST_TOPIC',
+  BROADCAST_THREAD: 'BROADCAST_THREAD',
   VECTORIZE_TOPIC: 'VECTORIZE_TOPIC',
+  VECTORIZE_THREAD: 'VECTORIZE_THREAD',
 } as const);
